删除集群时同步清理集群同步状态 #3226
Former-commit-id: 910aa8c934a774d90e5bf1e4304d1cc78b86d04a

+++ src/scene_server/topo_server/core/operation/inst_set.go
@@ -188,6 +188,12 @@ func (s *set) DeleteSet(params types.ContextParams, setModel model.Object, bizID
 		return params.Err.New(common.CCErrTopoSetDeleteFailed, err.Error())
 	}
 
+	// clear set template sync status
+	if ccErr := s.clientSet.CoreService().SetTemplate().DeleteSetTemplateSyncStatus(params.Context, params.Header, bizID, setIDS); ccErr != nil {
+		blog.Errorf("[operation-set] failed to delete set template sync status failed, bizID: %d, setIDs: %+v, err: %s, rid: %s", bizID, setIDS, err.Error(), params.ReqID)
+		return ccErr
+	}
+
 	// clear the sets
 	return s.inst.DeleteInst(params, setModel, setCond, false)
 }
